+++ CHANGELOG.md
@@ -4,10 +4,16 @@
 
 | Laravel Version   | Package Version   | Branch           |
 | ----------------- | ----------------- |------------------|
-| v8                | 4.x               | master           |
+| v9                | 5.x               | master           |
+| v8                | 4.x               | 4.x              |
 | v7                | 3.x               | version/v3.x     |
 | v6                | 2.x               |
 | v5.8              | 1.1.1             |
+
+## v5.x 
+- Support for Laravel 9
+- Drop Php 7 Support
+- Replace `cviebrock/eloquent-sluggable` with  `spatie/laravel-sluggable`
 
 ## v4.x
 - Support for Laravel 8